Revert to multi-line scanning
We need to allow line breaks inside the '{}'.  But to extract the
\BaseDirectory and \usedir commands _in the proper order_, we need to
scan for all macros simultaneously.  This makes for a rather complicated
regular expression, but it appears to work.  I suspect it could be
simplified, but I don't see off hand how to do that and still keep the
loop logic straight forward.