Clean up the index and glossary generation
The index and glossary generation happen as part of the PDF build.  The
PDF Builder does not look for outside builders but rather uses its
internal methods.  That means we have to use the variables to change the
behavior of the steps to build the index and glossary (for the change
log).  We drop the Builders and set default values for the
MAKEINDEXFLAGS and MAKEGLOSSARYSTYLE.  This does overwrite the global
default.  We may want to defer this to the user so it does not change
the default behavior of the main PDF builder.